Create clean architecture packages and move MainActivity to presentation/movies
