+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.udacity.popmoviesnd"
-          xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest
+    package="com.udacity.popmoviesnd"
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
         android:allowBackup="true"
@@ -9,7 +10,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+
+        <activity android:name=".presentation.movies.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
