+++ app/src/main/java/com/udacity/popmoviesnd/domain/model/Movie.java
@@ -1,0 +1,8 @@
+package com.udacity.popmoviesnd.domain.model;
+
+/**
+ * Class representing a Movie
+ */
+public final class Movie {
+
+}
