+++ app/src/main/java/com/udacity/popmoviesnd/presentation/movies/MainActivity.java
@@ -1,7 +1,9 @@
-package com.udacity.popmoviesnd;
+package com.udacity.popmoviesnd.presentation.movies;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.udacity.popmoviesnd.R;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -10,4 +12,5 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
     }
+
 }
